Split weighted functions out into own file.
